chore: move things around more
