+++ server/server.ts
@@ -5,7 +5,7 @@ import compression from 'compression';
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
 import enforce from 'express-sslify';
-import express, { RequestHandler } from 'express';
+import express, { ErrorRequestHandler, RequestHandler } from 'express';
 import noSlash from 'no-slash';
 import passport from 'passport';
 import path from 'path';
@@ -57,10 +57,26 @@ export const wrap: Wrap =
 		});
 	};
 
+const errorHandler: ErrorRequestHandler = (err, req, res, next) => {
+	if (err.message.indexOf('UseCustomDomain:') === 0) {
+		const customDomain = err.message.split(':')[1];
+		return res.redirect(`https://${customDomain}${req.originalUrl}`);
+	}
+	// Log the error if we're testing. Normally this is handled in the error middleware, but
+	// that isn't active while handling individual requests in a test environment.
+	if (process.env.NODE_ENV === 'test' && !(err instanceof HTTPStatusError)) {
+		// eslint-disable-next-line no-console
+		console.log('Got an error in an API route while testing:', err);
+	}
+	return next(err);
+};
+
 /* ---------------------- */
 /* Initialize express app */
 /* ---------------------- */
 const app = express();
+app.use(errorHandler);
+
 export default app;
 
 if (process.env.NODE_ENV === 'production') {
